brain_python/requirements.txt
grpcio==1.78.0
grpcio-tools==1.78.0
protobuf==4.25.1
colorama
requests
keyboard
sounddevice
numpy
ollama
pystray
Pillow
